waybar: fix version and switch to howard-hinnant-date
howard-hinnant-date has been upstreamed, so no reason to carry this
around anymore. Also, apply the sway-version treatment to make it
similar to building from git. Since we don't keep the .git folder,
remove the `git` build dependency in order to use our metadata as the
version. The format is slightly different than if we were to build from
source, but it is better than nothing:

    Waybar v (Jan  1 1970, branch '') # ours (old)

vs

    Waybar v0.9.1-5db06d99 (Mar 15 2020, branch 'master') # ours (new)

vs

    Waybar v0.9.0-44-ga555a72 (Mar  2 2020, branch 'master') # from source