auto-update(manual): xdg-desktop-portal-wlr: e103e120e20e83acec5b66ce4cd33a0eb86df0b0 => ab8ff54f4c3e564595afefbcbdd6c87e2c8188f5
Prior to this change, this package was pinned due to newer commits
referencing library: iniparser which is not distributed with a
pkg-config causing the build to fail on nix systems.

This change hacks around the problem by doing a replacement in
meson.build to cc.findlibrary(..) on the iniparser location in the nix
store.

Fixes: #283